Removed object overlay helper from object detection
